requirements.txt
livestreamer==1.12.2
av==0.2.4
pytesseract==0.1.6
numpy==1.10.2
pillow==3.0.0
progressbar==2.3
scikit-learn==0.17
scipy==0.16.1
